Simplify code now it's possible
